Setup Chatclient in Handlers

+++ Src/Modules/AiOrchestration/Exceptions/TokenCountException.cs
@@ -4,7 +4,7 @@ namespace AiOrchestration.Exceptions;
 
 public class TokenCountException : DomainException
 {
-    public TokenCountException(int token_count)
+    public TokenCountException(long token_count)
         : base($"token_count: '{token_count}' is invalid.")
     {
     }
