+++ Src/Modules/AutoComplete/Autocomplete/Features/GenerateAutoComplete/V1/Request/Request.cs
@@ -1,4 +1,4 @@
 ﻿namespace AutoComplete.Features.GenerateAutoComplete.V1;
 
 public record GenerateAutoCompleteRequestDto(string Prompt);
-public record GenerateAutoCompleteResponseDto(string Completion, int TokensUsed, decimal EstimatedCost);
+public record GenerateAutoCompleteResponseDto(string Completion, long TokensUsed, decimal EstimatedCost);
